Apply batch-two review fixes and document legacy-release IMU patch path
Code correctness (tracked in docs/code_review_findings_20260820.md):
- rotary attention: invert bool masks before SDPA (MHA True=masked vs SDPA
  True=allowed); causal decoding leaked future tokens and NaN'd the last
  position. Regression tests cover leakage, padding isolation, finiteness.
- multitask pretraining: heads lacked _target_ (dicts crashed on first
  forward) and tds_no_out's first conv expected 8ch against the harmonized
  16ch pretraining corpus; config-level tests instantiate and forward the
  module with 16ch input.
- vision validity on missing precomputed crops now False (fast-path parity);
  batch MixUp blends label_valid_mask with the same coefficient; ResNet
  head_vision keys no longer double-stripped; torch.load sites consistently
  pass weights_only=False; center-frame eval resolves split ids by manifest
  labels, runs on CPU, and takes a configurable EMG preference; keystroke
  collate probes the whole batch with None-safe consumers; evaluation sets
  EGOEMG_NO_PRETRAINED_DOWNLOAD to avoid ImageNet/DINO downloads offline;
  pooled evaluation pairs datasets by (split, hand) key; numpy augmentations
  preserve float32; missing EMG variants fail at construction with a
  config-pointing message. Center-frame black-frame inclusion kept by
  maintainer decision (documented).

Legacy-release IMU data readiness:
- the local release view already contains the fixed unified memmap (symlink);
  the 3.2 GB pre-fix backup moved out of the release view; ASSET_SETUP gains
  section 7 with post-fix SHA-256s, an in-place repair command for
  pre-2026-08-20 downloads, and a BaiduPCS-Go runbook to patch the cloud
  package by replacing only imu.dat + manifest.json; README release status
  documents the IMU field layout and the repair path.

Verified: 54 tests pass, flake8 clean, doc references resolve, fix-script
idempotence guard refuses on already-fixed data.

+++ egoemg/transforms.py
@@ -900,7 +900,7 @@ class BaselineDrift:
         else:
             scale = np.full((1, 1), np.std(x))
 
-        drift = np.zeros_like(x, dtype=float)
+        drift = np.zeros_like(x, dtype=x.dtype)
         if use_sine:
             freq = np.random.uniform(self.min_freq, self.max_freq)
             phase = np.random.uniform(
@@ -968,7 +968,7 @@ class PowerlineNoise:
         else:
             scale = np.full((1, 1), np.std(x))
 
-        noise = np.zeros_like(x, dtype=float)
+        noise = np.zeros_like(x, dtype=x.dtype)
         for harmonic in range(1, self.max_harmonic + 1):
             freq = base * harmonic
             if freq >= nyquist:
